Implement support for circuit breakers.
Extends the throttler with minimal resiliency capabilities. The
throttler now can receive a `:circuit_breaker` object responding to
`success`, `failure`, and `allow_request?`.

When an error occurrs in the throttler, this sends
the `failure` message to the circuit breaker object;
if it succeeds, it sends the `success` message.

With the information provided, the circuit breaker decides whether or
not to `allow_request?` the next time
the `throttle` checks freno, raising a `CircuitOpen` error in case it
does not allow the next call.

By default `Throttler::CircuitBreaker::Noop` is used in the throttler,
which does nothing and works as if no circuit breaker existed.